Update DoubleSliderPage.xaml.cs double zu int :D

+++ MobileDataCollection.Survey/MobileDataCollection.Survey/Views/DoubleSliderPage.xaml.cs
@@ -18,12 +18,12 @@ namespace MobileDataCollection.Survey.Views
 		}
 		 void OnSliderAValueChanged(object sender, ValueChangedEventArgs args)
         {
-            double value = args.NewValue;
+            int value = (int)(args.NewValue);
             sliderALabel.Text = String.Format("(A): {0}%", value);
         }
         void OnSliderBValueChanged(object sender, ValueChangedEventArgs args)
         {
-            double value = args.NewValue;
+            int value = (int)(args.NewValue);
             sliderBLabel.Text = String.Format("(B): {0}%", value);
         }
         void OnWeiterButtonClicked(object sender, EventArgs e)
